+++ rog-slash/src/data.rs
@@ -12,7 +12,7 @@ use crate::error::SlashError;
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Deserialize, Serialize)]
 pub enum SlashType {
     GA403,
-    Unknown,
+    GA605,
     Unsupported,
 }
 
@@ -22,7 +22,8 @@ impl FromStr for SlashType {
     fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         Ok(match s {
             "ga403" | "GA403" => Self::GA403,
-            _ => Self::Unknown,
+            "ga605" | "GA605" => Self::GA605,
+            _ => Self::Unsupported,
         })
     }
 }